backend/python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-genai>=1.0.0
pydantic==2.10.0
python-multipart==0.0.12

